Enable a faster mod in a couple places
